Move smart accessors out of Client.Sandbox into Client.Setup
